requirements.txt
# base
django>=1.5
django-rq
djangorestframework
django-filter
redis
pytz

# build
pillow
psycopg2

# develop
django-debug-toolbar
fabric
mock
sphinx